Refactor gym and quest popups
