+++ src/app/gym-popup/gym-popup.component.html
@@ -1,1 +1,57 @@
-<p>gym-popup works!</p>
+<ng-container *ngIf="isGymEdit then gymEditBlock else gymBlock"></ng-container>
+
+<ng-template #gymBlock>
+  <h1 mat-dialog-title class="gym-name-header">{{gymData.name}}<span *ngIf="gymData.isLegacy">*</span>
+    <button mat-icon-button (click)="toggleEditGym()">
+      <mat-icon>create</mat-icon>
+    </button>
+  </h1>
+  <div mat-dialog-content>
+    <div class="gym-pic">
+      <div id="swiper" class="union">
+        <img class="gym-under" src="//images.weserv.nl/?url={{gymData.imageUrl}}&w=126&h=174" [alt]="gymData.name" appFallback />
+        <img class="gym-over" src="../../assets/gyms/GymBadgeFrame_{{gymData.badge}}.png" />
+        <button mat-icon-button class="lefter" (click)="downgrade()" [disabled]="gymData.badge == 0 || gymData.isLegacy"><mat-icon>chevron_left</mat-icon></button>
+        <button mat-icon-button class="righter" (click)="upgrade()" [disabled]="gymData.badge == last || gymData.isLegacy"><mat-icon>chevron_right</mat-icon></button>
+      </div>
+    </div>
+
+  </div>
+  <div mat-dialog-actions>
+    <button mat-button (click)="setBadge()" [disabled]="oldBadge == gymData.badge">Save</button>
+    <button mat-button [mat-dialog-close]="gymData" cdkFocusInitial>Close</button>
+  </div>
+</ng-template>
+
+<ng-template #gymEditBlock>
+
+  <h1 mat-dialog-title>Update this gym</h1>
+
+  <div mat-dialog-content>
+
+    <form autocomplete="off" [formGroup]="gymUpdate" style="width: 300px;">
+      <mat-form-field>
+        <input matInput placeholder="Update the gym name" formControlName="name" required>
+        <mat-error *ngIf="f.name.errors">Not a valid name</mat-error>
+      </mat-form-field><br/>
+    
+      <mat-form-field>
+        <input matInput placeholder="Update the position" formControlName="pos" required>
+        <mat-error *ngIf="f.pos.errors">Not a valid position</mat-error>
+      </mat-form-field><br/>
+    
+      <mat-form-field>
+        <input matInput placeholder="Update the url" formControlName="imageUrl" required>
+        <mat-error *ngIf="f.imageUrl.errors">Not a valid image url</mat-error>
+      </mat-form-field><br/>
+  
+      <mat-checkbox color="primary" formControlName="isLegacy">Is legacy</mat-checkbox>
+    </form>
+  </div>
+
+  <div mat-dialog-actions>
+    <button mat-button [disabled]="gymUpdate.status != 'VALID' && !gymUpdate.dirty" (click)="saveGymEdit()">Update</button>
+    <button mat-button (click)="toggleEditGym()" cdkFocusInitial>Back</button>
+  </div>
+
+</ng-template>
