+++ src/app/gym-popup/gym-popup.component.scss
@@ -1,0 +1,110 @@
+$phoneBreakpoint: 600px;
+
+@mixin phone {
+    @media (max-width: #{$phoneBreakpoint}) {
+        @content;
+    }
+}
+
+.gym-name-header {
+    width: 300px;
+    overflow-wrap: normal;
+}
+
+.gym-pic {
+
+    position: relative;
+    height: 320px;
+    width: 320px;
+
+    @include phone() {
+      
+        height: 260px;
+        width: 260px;
+    }
+
+    .union {
+
+        position: absolute;
+        width: 256px;
+        height: 256px;
+        left: 50%;
+        top: 50%;
+        margin-top: -128px;
+        margin-left: -128px;
+
+        @include phone() {
+            height: 200px;
+            width: 200px;
+            margin-top: -100px;
+            margin-left: -100px;
+        }
+    }
+
+    .lefter {
+
+        margin-top: 116px;
+        margin-left: 25px;
+
+        @include phone() {
+            margin-top: 82px;
+            margin-left: 17px;
+        }
+    }
+
+    .righter {
+
+        margin-top: 116px;
+        margin-right: 25px;
+        float: right;
+
+        @include phone() {
+            margin-top: 82px;
+            margin-right: 17px;
+        }
+    }
+
+    img {
+
+        position: absolute;
+        -webkit-user-drag: none;
+    }
+
+    img.gym-under {
+
+        clip-path: polygon(50% 0%, 100% 50%, 50% 100%, 0% 50%);
+        width: 136px;
+        height: 176px;
+        margin: 40px 60px;
+
+        @include phone() {
+            width: 96px;
+            height: 134px;
+            top: 50%;
+            left: 50%;
+            margin-top: -67px;
+            margin-left: -48px;
+        }
+    }
+
+    img.gym-over {
+
+        @include phone() {
+            width: 196px;
+            height: 196px;
+            top: 50%;
+            left: 50%;
+            margin-top: -98px;
+            margin-left: -98px;
+        }
+    }
+
+}
+
+.gym-name {
+    text-decoration: none
+}
+
+mat-form-field {
+    width: 90%;
+}
