+++ src/app/quest-popup/quest-popup.component.html
@@ -1,82 +1,23 @@
-<ng-container *ngIf="data.imageUrl; then (isGymEdit ? gymEditBlock : gymBlock) else questBlock"></ng-container>
+<h1 mat-dialog-title>{{questData.stopName}}</h1>
+<div mat-dialog-content>
+  <p class="task">{{questData.taskDesc}}</p><br />
+  <p class="reward" *ngIf="questData.quantity">{{questData.quantity}} {{questData.reward}}</p><br />
 
-<ng-template #gymBlock>
-  <h1 mat-dialog-title class="gym-name-header">{{data.name}}<span *ngIf="data.isLegacy">*</span>
-    <button mat-icon-button (click)="toggleEditGym()">
-      <mat-icon>create</mat-icon>
-    </button>
-  </h1>
-  <div mat-dialog-content>
-    <div class="gym-pic">
-      <div id="swiper" class="union">
-        <img class="gym-under" src="//images.weserv.nl/?url={{data.imageUrl}}&w=126&h=174" [alt]="data.name" appFallback />
-        <img class="gym-over" src="../../assets/gyms/GymBadgeFrame_{{data.badge}}.png" />
-        <button mat-icon-button class="lefter" (click)="downgrade()" [disabled]="data.badge == 0 || data.isLegacy"><mat-icon>chevron_left</mat-icon></button>
-        <button mat-icon-button class="righter" (click)="upgrade()" [disabled]="data.badge == last || data.isLegacy"><mat-icon>chevron_right</mat-icon></button>
-      </div>
-    </div>
+  <p class="reward" *ngIf="questData.encounter == 'multiple'">Possible: </p><br />
+  <img class="quest-img" src="../../assets/quests/{{questData.reward == '#' ? questData.encounter : questData.reward}}.png" /><br />
 
-  </div>
-  <div mat-dialog-actions>
-    <button mat-button (click)="setBadge()" [disabled]="oldBadge == data.badge">Save</button>
-    <button mat-button [mat-dialog-close]="data" cdkFocusInitial>Close</button>
-  </div>
-</ng-template>
+  <label id="status-radio-group-label">Set quest status</label>
+  <mat-radio-group aria-labelledby="status-radio-group-label" class="status-radio-group" [(ngModel)]="questData.status">
+    <mat-radio-button class="status-radio-button" value="0">Not done</mat-radio-button>
+    <mat-radio-button class="status-radio-button" value="1">In progress</mat-radio-button>
+    <mat-radio-button class="status-radio-button" value="2">Done</mat-radio-button>
+  </mat-radio-group>
 
-<ng-template #gymEditBlock>
-
-  <h1 mat-dialog-title>Update this gym</h1>
-
-  <div mat-dialog-content>
-
-    <form autocomplete="off" [formGroup]="gymUpdate" style="width: 300px;">
-      <mat-form-field>
-        <input matInput placeholder="Update the gym name" formControlName="name" required>
-        <mat-error *ngIf="f.name.errors">Not a valid name</mat-error>
-      </mat-form-field><br/>
-    
-      <mat-form-field>
-        <input matInput placeholder="Update the position" formControlName="pos" required>
-        <mat-error *ngIf="f.pos.errors">Not a valid position</mat-error>
-      </mat-form-field><br/>
-    
-      <mat-form-field>
-        <input matInput placeholder="Update the url" formControlName="imageUrl" required>
-        <mat-error *ngIf="f.imageUrl.errors">Not a valid image url</mat-error>
-      </mat-form-field><br/>
-  
-      <mat-checkbox color="primary" formControlName="isLegacy">Is legacy</mat-checkbox>
-    </form>
-  </div>
-
-  <div mat-dialog-actions>
-    <button mat-button [disabled]="gymUpdate.status != 'VALID' && !gymUpdate.dirty" (click)="saveGymEdit()">Update</button>
-    <button mat-button (click)="toggleEditGym()" cdkFocusInitial>Back</button>
-  </div>
-
-</ng-template>
-
-<ng-template #questBlock>
-  <h1 mat-dialog-title>{{data.stopName}}</h1>
-  <div mat-dialog-content>
-    <p class="task">{{data.taskDesc}}</p><br />
-    <p class="reward" *ngIf="data.quantity">{{data.quantity}} {{data.reward}}</p><br />
-
-    <p class="reward" *ngIf="data.encounter == 'multiple'">Possible: </p><br />
-    <img class="quest-img" src="../../assets/quests/{{data.reward == '#' ? data.encounter : data.reward}}.png" /><br />
-
-    <label id="status-radio-group-label">Set quest status</label>
-    <mat-radio-group aria-labelledby="status-radio-group-label" class="status-radio-group" [(ngModel)]="data.status">
-      <mat-radio-button class="status-radio-button" value="0">Not done</mat-radio-button>
-      <mat-radio-button class="status-radio-button" value="1">In progress</mat-radio-button>
-      <mat-radio-button class="status-radio-button" value="2">Done</mat-radio-button>
-    </mat-radio-group>
-
-  </div>
-  <div mat-dialog-actions>
-    <button mat-button (click)="exclude(0)">Exclude this type</button>
-    <button mat-button (click)="exclude(1)" *ngIf="data.encounter != 'multiple'">Exclude this {{data.reward == "#" ? "encounter" : "reward"}}</button>
-    <button mat-button (click)="setStatus()">Save</button>
-    <button mat-button [mat-dialog-close]="data" cdkFocusInitial>Close</button>
-  </div>
-</ng-template>
+</div>
+<div mat-dialog-actions>
+  <button mat-button (click)="exclude(0)">Exclude this type</button>
+  <button mat-button (click)="exclude(1)" *ngIf="questData.encounter != 'multiple'">Exclude this
+    {{questData.reward == "#" ? "encounter" : "reward"}}</button>
+  <button mat-button (click)="setStatus()">Save</button>
+  <button mat-button [mat-dialog-close]="questData" cdkFocusInitial>Close</button>
+</div>
